vision_llm/requirements.txt
fastapi
uvicorn
psycopg2-binary
minio
numpy
torch
transformers
sentence-transformers
